feat: add Seoul risk map rendering
