+++ src/features/recovery-zone/index.ts
@@ -1,7 +1,1 @@
-// export { LayerToggle } from "./components/LayerToggle";
-// export { MapSidebar } from "./components/MapSidebar";
-// export { useSelectedLayer } from "./hooks/useSelectedLayer";
-
-// 이런식으로 하면 다른 곳에서 import할때 간결하게 가져올 수 있음!
-
-// import { LayerToggle, MapSidebar } from "@/features/sinkhole-map";
+export { default as MapSeoulRisk } from './ui/MapSeoulRisk';
